fixing bugs

+++ src/api/lesson/controllers/lesson.js
@@ -3,114 +3,78 @@
 const { createCoreController } = require("@strapi/strapi").factories;
 
 module.exports = createCoreController("api::lesson.lesson", ({ strapi }) => ({
-  async find(ctx) {
-    try {
-      const { _page = 1, _limit = 5, _q = "", moduleId } = ctx.query;
-      const page = parseInt(_page, 10);
-      const limit = parseInt(_limit, 10);
-      const start = (page - 1) * limit;
-
-      const where = {
-        nom: { $contains: _q },
-        module: moduleId,
-        users_permissions_user: {
-          id: ctx.state.user.id,
-        },
-      };
-
-      const [lessons, total] = await strapi.db
-        .query("api::lesson.lesson")
-        .findWithCount({
-          where,
-          start,
-          limit,
-          sort: { createdAt: "DESC" },
-        });
-
-      ctx.send({
-        data: lessons,
-        meta: {
-          pagination: {
-            page,
-            pageSize: limit,
-            pageCount: Math.ceil(total / limit),
-            total,
-          },
-        },
-      });
-    } catch (error) {
-      console.error("Erreur lors de la récupération des leçons :", error);
-      ctx.throw(500, "Erreur lors de la récupération des leçons");
-    }
-  },
-
   async create(ctx) {
     try {
-      const { name, module } = ctx.request.body;
+      const { data } = ctx.request.body;
 
-      // Vérifiez si le module existe
+      if (!data.nom || !data.module) {
+        return ctx.badRequest("Le nom de la leçon et le module sont requis");
+      }
+
       const existingModule = await strapi.entityService.findOne(
         "api::module.module",
-        module
+        data.module
       );
 
       if (!existingModule) {
         return ctx.throw(404, "Module non trouvé");
       }
 
-      // Créez la leçon
-      const createdLesson = await strapi.entityService.create(
+      const newLesson = await strapi.entityService.create(
         "api::lesson.lesson",
         {
           data: {
-            nom: name,
-            module: module,
+            nom: data.nom,
+            module: data.module,
             users_permissions_user: ctx.state.user.id,
-            publishedAt: new Date(), // Publier automatiquement la leçon
+            publishedAt: new Date(),
           },
         }
       );
 
-      ctx.send({
-        message: "Leçon créée avec succès",
-        data: createdLesson,
-      });
+      ctx.send({ message: "Leçon créée avec succès", data: newLesson });
     } catch (error) {
-      ctx.throw(500, "Erreur lors de la création de la leçon");
+      console.error("Erreur lors de la création de la leçon :", error);
+      ctx.throw(500, "Une erreur est survenue lors de la création de la leçon");
+    }
+  },
+
+  async find(ctx) {
+    try {
+      const { _page = 1, _limit = 5, _q = "", module } = ctx.query;
+      const page = parseInt(_page, 10) || 1;
+      const limit = parseInt(_limit, 10) || 5;
+      const query = _q || "";
+      const userId = ctx.state.user.id;
+
+      const result = await strapi
+        .service("api::lesson.lesson")
+        .findLessons(page, limit, query, module, userId);
+
+      ctx.send(result);
+    } catch (error) {
+      console.error("Erreur lors de la récupération des leçons:", error);
+      ctx.throw(
+        500,
+        "Une erreur est survenue lors de la récupération des leçons"
+      );
     }
   },
 
   async update(ctx) {
     try {
       const { id } = ctx.params;
-      const { data } = ctx.request.body;
+      const { nom } = ctx.request.body;
 
-      const updatedLesson = await strapi.entityService.update(
-        "api::lesson.lesson",
-        id,
-        {
-          data: {
-            nom: data.nom,
-            users_permissions_user: ctx.state.user.id,
-            publishedAt: new Date(), // Publier automatiquement la leçon mise à jour
-          },
-        }
-      );
+      console.log("====================================");
+      console.log("this is  puttt data ");
+      console.log(ctx.request.body);
 
-      ctx.send({
-        message: "Leçon mise à jour avec succès",
-        data: updatedLesson,
-      });
-    } catch (error) {
-      ctx.throw(500, "Erreur lors de la mise à jour de la leçon");
-    }
-  },
+      console.log("====================================");
+      if (!nom) {
+        return ctx.badRequest("Le nom de la leçon est requis");
+      }
 
-  async delete(ctx) {
-    try {
-      const { id } = ctx.params;
-
-      // Vérifiez si la leçon existe
       const existingLesson = await strapi.entityService.findOne(
         "api::lesson.lesson",
         id
@@ -120,14 +84,84 @@ module.exports = createCoreController("api::lesson.lesson", ({ strapi }) => ({
         return ctx.throw(404, "Leçon non trouvée");
       }
 
-      // Supprimez la leçon
+      const updatedLesson = await strapi.entityService.update(
+        "api::lesson.lesson",
+        id,
+        {
+          data: {
+            nom,
+            users_permissions_user: ctx.state.user.id,
+            publishedAt: new Date(),
+          },
+        }
+      );
+
+      ctx.send({
+        message: "Leçon mise à jour avec succès",
+        data: updatedLesson,
+      });
+    } catch (error) {
+      console.error("Erreur lors de la mise à jour de la leçon :", error);
+      ctx.throw(
+        500,
+        "Une erreur est survenue lors de la mise à jour de la leçon"
+      );
+    }
+  },
+
+  async delete(ctx) {
+    try {
+      const { id } = ctx.params;
+
+      const existingLesson = await strapi.entityService.findOne(
+        "api::lesson.lesson",
+        id
+      );
+
+      if (!existingLesson) {
+        return ctx.throw(404, "Leçon non trouvée");
+      }
+
       await strapi.entityService.delete("api::lesson.lesson", id);
 
       ctx.send({
         message: "Leçon supprimée avec succès",
       });
     } catch (error) {
-      ctx.throw(500, "Erreur lors de la suppression de la leçon");
+      console.error("Erreur lors de la suppression de la leçon :", error);
+      ctx.throw(
+        500,
+        "Une erreur est survenue lors de la suppression de la leçon"
+      );
+    }
+  },
+  async findOne(ctx) {
+    try {
+      const { id } = ctx.params;
+      const userId = ctx.state.user.id;
+
+      const existingLesson = await strapi.entityService.findOne(
+        "api::lesson.lesson",
+        id,
+        {
+          filters: { users_permissions_user: userId },
+        }
+      );
+
+      if (!existingLesson) {
+        return ctx.throw(404, "Leçon non trouvée");
+      }
+
+      ctx.send({
+        message: "Leçon récupérée avec succès",
+        data: existingLesson,
+      });
+    } catch (error) {
+      console.error("Erreur lors de la récupération de la leçon :", error);
+      ctx.throw(
+        500,
+        "Une erreur est survenue lors de la récupération de la leçon"
+      );
     }
   },
 }));
